Update container script

+++ replace-container.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-docker stop con
-docker system prune -f
-docker pull ziwei0/awsdevops
-docker run -d -p 8080:8080 --name con ziwei0/awsdevops
+if ! docker pull ziwei0/awsdevops | grep "Image is up to date" ; then
+    docker stop con
+    docker system prune -f
+    docker run -d -p 8080:8080 --name con ziwei0/awsdevops
+fi
